improved onboarding and navbar
